refactor: 모달 버튼 디렉토리 위치 변경
