+++ src/components/Modal/ModalButtons/StopSpeechBtn.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Stop } from 'src/utils/Icons';
-import * as S from './Styles';
+import * as S from '../Styles';
 
 interface Props {
   onClick: () => void;
