+++ src/components/Modal/ModalResult.tsx
@@ -5,8 +5,8 @@ import { incorrectWordsSelector } from 'src/recoil/selectors';
 import { audioUrlState, recognizedSentence, recordingState, scriptSentencestate } from 'src/recoil/states';
 import styled from 'styled-components';
 import { Checkmark } from 'react-checkmark';
-import PlaySpeechBtn from './PlaySpeechBtn';
-import ReSpeechBtn from './ReSpeechBtn';
+import PlaySpeechBtn from './ModalButtons/PlaySpeechBtn';
+import ReSpeechBtn from './ModalButtons/ReSpeechBtn';
 
 function ModalResult() {
   const [isPlay, setIsPlay] = useState(false);
